Organize all the webhook parser unit tests into a folder
