+++ Source/ZoomNet.UnitTests/WebhookParser/ContactCenterWebhookParserTests.cs
@@ -1,11 +1,11 @@
-using System;
 using Shouldly;
+using System;
 using Xunit;
 using ZoomNet.Models;
 using ZoomNet.Models.Webhooks;
 using ZoomNet.UnitTests.Properties;
 
-namespace ZoomNet.UnitTests
+namespace ZoomNet.UnitTests.WebhookParser
 {
 	/// <summary>
 	/// Unit tests that verify contact center webhook events parsing.
